requirements.txt
web3 
python-socks
web3-proxy-providers
PyYaml
pycrypto
telebot
pandas
PyQt5